requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
flower==2.0.1
gunicorn==23.0.0
humanize==4.10.0
kombu==5.4.2
packaging==24.1
prometheus_client==0.21.0
prompt_toolkit==3.0.47
psycopg==3.2.2
psycopg-binary==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
redis-cli==1.0.1
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
